improve docstrings and reorganize code for clarity
