+++ xoto3/dynamodb/write_versioned/modify.py
@@ -1,4 +1,4 @@
-"""Defines the API for modifying an existing transaction. """
+"""Defines the API for adding write effects to a transaction."""
 
 from typing import NamedTuple, Optional, Union
 
@@ -10,8 +10,7 @@ from xoto3.utils.tree_map import SimpleTransform
 from .ddb_api import known_key_schema
 from .ddb_api import table_name as _table_name
 from .errors import TableSchemaUnknownError
-from .keys import hashable_key, key_from_item
-from .prepare import standard_key_attributes_from_key
+from .keys import hashable_key, key_from_item, standard_key_attributes
 from .types import TableNameOrResource, VersionedTransaction, _TableData
 
 
@@ -33,14 +32,7 @@ def _write(
     *,
     nicename: str = DEFAULT_ITEM_NAME,
 ) -> VersionedTransaction:
-    """Shared put/delete implementation - not meant for direct use at this time.
-
-    Performs an optimistic put - if the item is not known to the
-    existing transaction, assumes you mean to create it if and only if
-    it does not exist. If the item turns out to exist already, your
-    transaction will be re-run, at which point a put will be interpreted as a
-    'witting' choice to overwrite the known item.
-    """
+    """Shared put/delete implementation - not meant for direct use at this time."""
 
     table_name = _table_name(table)
 
@@ -63,7 +55,7 @@ def _write(
                         "Specify this delete in terms of the key only and this should work fine."
                     )
                 # at this point this is a best guess
-                key_attributes = standard_key_attributes_from_key(put_or_delete.item_key)
+                key_attributes = standard_key_attributes(*put_or_delete.item_key.keys())
             else:
                 # it's a put - we can't make this work at all
                 raise TableSchemaUnknownError(
@@ -100,7 +92,14 @@ def put(
     nicename: str = DEFAULT_ITEM_NAME,
     prewrite_transform: Optional[SimpleTransform] = None,
 ) -> VersionedTransaction:
-    """Returns a modified transaction including the requested PutItem operation"""
+    """Returns a modified transaction including the requested PutItem operation
+
+    Performs an optimistic put - if the item is not known to the
+    existing transaction, assumes you mean to create it if and only if
+    it does not exist. If the item turns out to exist already, your
+    transaction will be re-run, at which point a put will be interpreted as a
+    'witting' choice to overwrite the known item.
+    """
     return _write(
         transaction, table, Put(dynamodb_prewrite(item, prewrite_transform)), nicename=nicename
     )
@@ -115,76 +114,3 @@ def delete(
 ) -> VersionedTransaction:
     """Returns a modified transaction including the requested DeleteItem operation"""
     return _write(transaction, table, Delete(item_or_key), nicename=nicename)
-
-
-def presume(
-    transaction: VersionedTransaction,
-    table: TableNameOrResource,
-    item_key: ItemKey,
-    item_value: Optional[Item],
-) -> VersionedTransaction:
-
-    """'To assume as true in the absence of proof to the contrary.'
-
-    Returns a modified transaction if the value of the item is not already
-    known. This will make the presumed value available via `get`,
-    and will additionally check your presumed value against the table when
-    the transaction is run.
-
-    This is purely a cost optimization to avoid fetching something
-    when we believe we already know its value. Any transaction builder
-    will result in exactly the same data written to the table with or
-    without this statement. If the item turns out to have a different
-    value in the table than you presumed when the transaction is
-    executed, the transaction will restart, the item will be freshly
-    fetched like usual, and this procedure will have no effect on the
-    transaction.
-
-    This may also be used for the purpose of stubbing out values in an
-    empty VersionedTransaction for writing unit tests against your
-    transaction builder functions.
-    """
-    if item_value is not None:
-        for key_attr, key_val in item_key.items():
-            assert item_value[key_attr] == key_val, "Item key must match in a non-nil item value"
-    table_name = _table_name(table)
-    if table_name in transaction.tables:
-        table_data = transaction.tables[table_name]
-    else:
-        table_data = _TableData(
-            items=dict(), effects=dict(), key_attributes=standard_key_attributes_from_key(item_key)
-        )
-    hkey = hashable_key(item_key)
-    if hkey not in table_data.items:
-        return VersionedTransaction(
-            tables={
-                **transaction.tables,
-                table_name: _TableData(
-                    items={**table_data.items, hkey: item_value},
-                    effects=table_data.effects,
-                    key_attributes=table_data.key_attributes,
-                ),
-            }
-        )
-    return transaction
-
-
-def define_table(
-    transaction: VersionedTransaction, table: TableNameOrResource, *key_attributes: str,
-) -> VersionedTransaction:
-    """
-    This is a convenient way of dynamically defining a key
-    attribute schema for a given table without forcing any IO
-    operations/effects up front.
-    """
-    assert len(key_attributes) > 0 and len(key_attributes) <= 2
-    if _table_name(table) in transaction.tables:
-        return transaction
-    return VersionedTransaction(
-        tables={
-            **transaction.tables,
-            _table_name(table): _TableData(
-                items=dict(), effects=dict(), key_attributes=tuple(sorted(key_attributes)),
-            ),
-        }
-    )
